+++ jsconfig.json
@@ -1,0 +1,7 @@
+{
+    "typeAcquisition": {
+        "include": [
+            "jest"
+        ]
+    }
+}
