+++ src/App.jsx
@@ -1,26 +1,18 @@
 import React, { useState, useEffect } from 'react';
-import './App.css';
-import ArrayBit from './components/ArrayBit/ArrayBit'
-import Button from './components/Button/Button'
 import { Slider, MenuItem, Select } from '@mui/material';
+import ArrayContainer from './components/ArrayContainer/ArrayContainer';
+import Button from './components/Button/Button'
+import handleSort from './action/Sort'
+import './App.css';
 
-function App() {
+const App = () => {
 
     const [r_array, setRArray] = useState([]);
     const [rangeValue, setRangeValue] = useState(35);
     const [isRunning, setIsRunning] = useState(false);
     const [sortMethod, setSortMethod] = useState(0);
     const [steps, setSteps] = useState(0);
-    const sliderLabel = [
-        {
-            value: 20,
-            label: 'Fast'
-        },
-        {
-            value: 200,
-            label: 'Slow'
-        }
-    ]
+    
     const createUnsortedArray = () => {
         setSteps(0)
         for (var array = [], i = 0; i < rangeValue; ++i) array[i] = i;
@@ -38,44 +30,6 @@ function App() {
         setRArray(array);
     };
 
-    const delay = () => {
-        return new Promise(
-            resolve => setTimeout(resolve, 0.00001)
-        );
-    }
-
-    const BubbleSort = async () => {
-        let arr = [...r_array]
-        let step = 0
-        for (let i = 0; i < arr.length; i++) {
-            for (let j = 0; j < arr.length; j++) {
-                await delay()
-                if (arr[j] > arr[j + 1]) {
-                    step++
-                    setSteps(step)
-                    let temp = arr[j];
-                    arr[j] = arr[j + 1];
-                    arr[j + 1] = temp;
-                    setRArray([...arr]);
-                }
-            }
-        }
-        console.log('done')
-        setIsRunning(false)
-    }
-
-    const handleSort = () => {
-        setIsRunning(true)
-        switch (sortMethod) {
-            case 0:
-                BubbleSort()
-                break;
-
-            default:
-                break;
-        }
-    }
-
     const handleSelectSort = (event) => {
         setSortMethod(event.target.value)
     }
@@ -88,13 +42,15 @@ function App() {
         setRangeValue(newValue)
     }
 
+    const handleSortLaunch = () => {
+        handleSort(sortMethod, r_array, setRArray, setIsRunning, setSteps)
+    }
+
     useEffect(() => {
         if (r_array.length === 0) {
             createUnsortedArray()
         }
     }, [r_array]);
-
-    // console.log('rendering');
 
     return (
         <div className="App" >
@@ -132,29 +88,13 @@ function App() {
                                 <MenuItem value={0}>Bubble Sort</MenuItem>
                             </Select>
                         <div className="button-container">
-                            <Button onClickAction={handleSort} desactivated={isRunning} >Launch Sort</Button>
+                            <Button onClickAction={handleSortLaunch} desactivated={isRunning} >Launch Sort</Button>
                         </div>
                         </div>
                     </div>
                 </div >
             </header>
-            <main>
-                <div className="arrayHolder" > {
-                    r_array.length !== 0 ? (
-
-                        r_array.map((val) => {
-                            return <ArrayBit key={val}
-                                index={val}
-                                arrayLength={r_array.length}
-                            />
-                        })
-                    ) : < div className='center-loading' > Loading... </div>
-                }
-                </div>
-            </main >
-            <div>
-                Steps: {steps}
-            </div>
+            <ArrayContainer targetArray={r_array} steps={steps}/>
         </div>
     );
 }
